# docs/swagger.yaml
definitions:
  request.HelloRequest:
    properties:
      age:
        example: 27
        type: integer
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
  request.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  request.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password
        maxLength: 100
        minLength: 6
        type: string
      password_confirmation:
        example: password
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    - password_confirmation
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        example: Message success
        type: string
      status:
        example: 200
        type: integer
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: Hello Handler
      parameters:
      - description: search by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response so the data field is array of any type
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Show a hello message
      tags:
      - Hello
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: json type
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success response so the data field is array of any type
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: json type
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success response so the data field is array of any type
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      tags:
      - Auth
  /post-hello:
    post:
      consumes:
      - application/json
      description: Hello message json body
      parameters:
      - description: json type
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.HelloRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success response so the data field is array of any type
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Post a hello message using json body
      tags:
      - Hello
  /post-hello-form:
    post:
      consumes:
      - multipart/form-data
      description: Hello message form data
      parameters:
      - description: Name
        in: formData
        name: name
        required: true
        type: string
      - description: Age
        in: formData
        name: age
        required: true
        type: integer
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: File to upload
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: success response so the data field is array of any type
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Post a hello message using form data
      tags:
      - Hello
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"